Moved different data types to new classes
