+++ Assets/Wiimote/Scripts/Enums.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 11a09387e01eab1418eac5fcf8ba40ed
+timeCreated: 1429997857
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
