+++ Assets/Wiimote/Scripts/Internal.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 36c2aa32fe85e404baa4fbfee1260b16
+folderAsset: yes
+timeCreated: 1430001740
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
