+++ Assets/Wiimote/Scripts/Internal/RegisterReadData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 17d61abebb90a8c49ae13f8559b942e9
+timeCreated: 1430001747
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
