+++ Assets/Wiimote/Scripts/Wiimote.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 5463cc53d9e56334aabf1ad0fe825524
+timeCreated: 1429997858
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
