+++ Assets/Wiimote/Scripts/WiimoteData.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: fae9f7dfa1e48f746955114d76a73ac7
+folderAsset: yes
+timeCreated: 1429998341
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
