+++ Assets/Wiimote/Scripts/WiimoteData/AccelData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: be22862c7701a934fa0a1559ab6f2cba
+timeCreated: 1429998422
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
