+++ Assets/Wiimote/Scripts/WiimoteData/ButtonData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 48c4286a80d7a3d4692e351e811cf7a0
+timeCreated: 1429998439
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
