+++ Assets/Wiimote/Scripts/WiimoteData/IRData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 8ce4c54d3c40d3246b360c5aa8314559
+timeCreated: 1429998414
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
