+++ Assets/Wiimote/Scripts/WiimoteData/MotionPlusData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: fcbc9f7491037954684ef4ecd777d906
+timeCreated: 1429998409
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
