+++ Assets/Wiimote/Scripts/WiimoteData/NunchuckData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0bbe453355b439742842ed55cd1f3924
+timeCreated: 1429997857
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
