+++ Assets/Wiimote/Scripts/WiimoteData/StatusData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 892e7813c76b70044b22b0d699412cbc
+timeCreated: 1430000300
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
