+++ Assets/Wiimote/Scripts/WiimoteData/WiimoteData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 60f7bcb0e4ca5984ca1186bd5c8af7db
+timeCreated: 1430000879
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
